Clipper2 (alpha) readme.txt correction to previous commit.
The C++ performance improvements reported in the previous commit were accurate in their comparisons with the older version of Clipper. However, there was no rounding of coordinates to 10 (as there is in the Delphi and C# benchmark tests) and so the C++ code appeared (incorrectly) to vastly out perform the Delphi and C# code. Rounding random coordinates to 10 severely stresses the clipping algorithm, and it's why I've been using it as a benchmark. Anyhow, there are significant performance improvements, they just aren't quite as noticeable due to the unusual stressed provided by this test. I've amended the C++ tests to show testing with coordinates rounded to 10.

git-svn-id: https://svn.code.sf.net/p/polyclipping/code/sandbox@531 758e7f8d-410f-429b-b4f6-457a56196ea1